Change setting \xff/minRequiredCommitVersion to not always cause a recovery
The recovery is now skipped if `version <= current version`.

There's been a couple new inqueries into using this key, and it turns
outout that the expected simple transaction body of just setting this
key causes an infinite number of recoveries, as `commit_unknown_result`
is a retryable error, and setting minRequiredCommitVersion would
previously never result in anything but commit_unknown_result.