FastRestore: Fix race condition  when same request is delivered twice
When the applyToDB request is delivered more than once before the old request is completely process,
the processing for the same type of request may cause race condition between
(the invocation of) the multiple actors for the same request.

We need to use a variable to identify if the worker is currently processing the request.
If it is, it needs to wait until the previous actor finishes processing the request.

TODO: The same race problem can happen to other requests as well.
We may want to make sure processing each request exactly once!!!