FastRestore: Add delay after each phase
This somehow makes sure the operation in the next phase will not
be executed in the current phase in the simulation, although this
should never be possible based on the code logic.

It seems that simulator may reorder the actor execution sequence
without checking some conditions?

After adding the wait(delay(1.0)), the race condition between actors
that causes corrupted memory/pointer disappears.