FastRestore: Fix race condition among the same actors
When the same actor is invoked multiple times, the logic
in the same actor will cause race condition, which may
lead to data corruption or inconsistent restored DB.

To solve this problem, we use inProgressFlag to record
the phase a worker is currently working on:
each set bit in inProgressFlag indicates a in-progress phase.
When a worker invoke an actor for a phase,
It first check the inProgressFlag if the worker is currently
working in the phase; if yes, the worker's actor for the same phase
will wait;
Once the actor is the only one working in the phase,
it sets the bit in inProgressFlag.
When the actor finishes the phase, it clears the bit.

This solve the race condition, although it does not perfectly solve it.
To solve the race condition, i.e., avoiding the rare race between machine instructions,
we need to use lock (or mutex) to ensure only one can access the flag.

This commit also enclose the unlockDB logic into a try-catch loop.
Otherwise, the unlockDB error will fail the correctness test.