StorageEngineSwitch:Fix out of memory bug
When there is at least one healthy team and there exists a
storage server with wrong storeType, we should pause recruiting logic.
Otherwise, recruiting too many on-the-fly servers will cause
out of memory bug.

We also need to ensure the storage server marked to be removed should be
removed in a timely manner.