StorageEngineSwitch:Reboot DD when config change
When config changes, we must reboot data distributor and let it
get the new config.
After data distributor was split from master as a separate role,
it was not guaranteed to reboot after config change.

Also fix a bug in storageRecruiter actor that does not exclude
addresses correctly, resulting keeping recruiting servers.