Choose LogRouter tag to match an existing tag to save a copy.
Let us suppose we have:
* five TLogs in the primary
* five log routers
* a mutation with tags {1,2,3}

Every tag has a preferred location (that is the id `mod` number of
logs).  For every tag on a mutation, the mutation *must* be stored at
the preferred location for that tag.

Previously, we would choose log router tags randomly, to ensure an even
distribution of mutations over log routers.  Suppose we happen to choose
-2:4 as the log router tag.  We already need to send the mutation to
TLogs 1, 2, and 3.  -2:4 has a preferred location of TLog 4, so now we
need to send out a fourth copy.

This commit changes the assignment to only choose from log router tags
that match an existing place where the mutation would already need to be
stored, to avoid the situation where a fourth copy would need to be
sent.  Within this example, we would only randomly choose over the
restricted subset of {-2:1, -2:2, -2:3} as the log router tags.

This potentially has the downside of being less efficient if one runs
one log router per primary TLog.  However, this is not generally the
recommended deployment, and Log Routers are stateless roles, which makes
running many of them very cheap and easy to do.