TeamTracker:BugFix:Must be triggered after server status change
Before this commit, we use a trigger in server_info to trigger the
teamTracker by the removeWrongStoreType actor.
The intention is to make sure whenever a server's storage type becomes invalid,
teamTracker is invoked to start a shard relocation request, which will move
data aways from the server.

The intention is correct except that teamTracker should be triggered after
the server's status changes, so that the number of servers left in the team
can be correctly counted.

Since the removeWrongStoreType actor also triggers the server_status
when a server storeType becomes invalid, we do not need to use
the server_info's trigger, which may cause the teamRemover invoked before
server status change.