FastRestore:Handle multiple files at the same version
When mutations at the same version can exist in backup files,
the current fast restore will get stuck because it implicitly assumes
this will not happen.

Since this situation can happen when new backup is introduced,
we should handle this situation.

In order to overcome this limitation, we use NotifiedVersion for each
file to keep track of the progress of receiving mutations on appliers.

By ensuring only one NotifiedVersion exists for a file, we ensure a file
is handled only once.
By using NotifiedVersion, we ensure mutations in a file are received by
applier only once.