Fix backup worker ID and remove some fields
The backup worker ID was changed to be the interface ID, not the request ID.
The lastSeenVersion is replaced with minKnownCommittedVersion, which is
incremented even when there is no mutations. This also avoid the problem of
setting popVersion to be higher than the actual committed version (no harm
here though, as there are no mutations).

The "backupStartedKey" handling is also fixed to correctly handle cases when
we wait for the stopping of backups.