FastRestore:A single version may be larger than version batch threshold
In case data at a single version is larger than FASTRESTORE_VERSIONBATCH_MAX_BYTES,
we should allow a version batch to include the version and ignore the
FASTRESTORE_VERSIONBATCH_MAX_BYTES limit to avoid false positive in simulation.

In real environment, this situation will report SevError to ask DBA to
increase the memory limit for a version batch.